+++ modules/ip_info/ip_info.py
@@ -212,8 +212,10 @@ class Module(Module, multiprocessing.Process):
         """
         if not hasattr(self, 'country_db'):
             return False
-
-        if geoinfo := self.country_db.get(ip):
+        if ipaddress.ip_address(ip).is_private:
+            # Try to find if it is a local/private IP
+            data = {'geocountry': 'Private'}
+        elif geoinfo := self.country_db.get(ip):
             try:
                 countrydata = geoinfo['country']
                 countryname = countrydata['names']['en']
@@ -221,9 +223,6 @@ class Module(Module, multiprocessing.Process):
             except KeyError:
                 data = {'geocountry': 'Unknown'}
 
-        elif ipaddress.ip_address(ip).is_private:
-            # Try to find if it is a local/private IP
-            data = {'geocountry': 'Private'}
         else:
             data = {'geocountry': 'Unknown'}
         __database__.setInfoForIPs(ip, data)
@@ -297,7 +296,7 @@ class Module(Module, multiprocessing.Process):
 
             if oui in line:
                 line = json.loads(line)
-                vendor = line['companyName']
+                vendor = line['vendorName']
                 return vendor
 
     def get_vendor(self, mac_addr: str, host_name: str, profileid: str):